Creative inventory action into new packet system
