+++ src/main/kotlin/packets/PacketHandler.kt
@@ -284,24 +284,6 @@ class PacketHandler(
     }
 
     @PacketReceiver
-    fun onCreativeInventoryAction(packet: ClientCreativeInventoryActionPacket) {
-        val slotIdx = packet.slotIndex.toInt()
-        val stack = if(packet.slot.present) packet.slot.toItemStack() else null
-        client.player.inventory.set(slotIdx, stack)
-
-        if (slotIdx == client.player.selectedSlot + 36) sendHeldItemUpdate()
-        if (slotIdx == -1) { //drop item
-            if (stack != null && !stack.isAir) {
-                val vx = ((Math.random() - 0.5) * 0.2 * 8000).toInt().toShort()
-                val vy = (0.1 * 8000).toInt().toShort()
-                val vz = ((Math.random() - 0.5) * 0.2 * 8000).toInt().toShort()
-
-                dropItem(client.player.location.toBlockPosition(), stack.id, vx, vy, vz)
-            }
-        }
-    }
-
-    @PacketReceiver
     fun onPluginMessage(packet: ClientPluginMessagePacket) {
         when (packet.channel) {
             "minecraft:brand" -> {
@@ -955,22 +937,6 @@ class PacketHandler(
         for (otherPlayer in Bullet.players) {
             if (otherPlayer != player) {
                 otherPlayer.sendPacket(packet)
-            }
-        }
-    }
-
-    private fun sendHeldItemUpdate() {
-        val stack = client.player.getHeldItem()
-        val slotData = stack.toSlotData()
-
-        for (otherPlayer in Bullet.players) {
-            if (otherPlayer != client.player) {
-                otherPlayer.sendPacket(
-                    ServerEntityEquipmentPacket(
-                        client.player.entityID,
-                        listOf(0 to slotData)
-                    )
-                )
             }
         }
     }
